Fix: Avoid race condition on debugging protocol
